add obisindicators tool
This tool is for generating indicators from OBIS.

I could not add DggridR in conda because the version needed, 3.0.0, is not yet available on CRAN (should be soon). Thus, for now I install dggridr with install_github. I don't know if we can keep it that way until the version 3.0.0 is on the cran or just we prepare the PR and keep it until I can add the package on conda ? (I hope I am clear enough)